display skeleton
