+++ src/app/api/posts/route.ts
@@ -1,9 +1,12 @@
 import {NextRequest, NextResponse} from "next/server";
 import {postSearch} from "@/app/dashboard/posts/_lib/post.search";
 import {searchHomePosts} from "@/app/(landing)/posts/_lib/post.queries";
+import {sleep} from "@/lib/sleep";
 
 export async function GET(req: NextRequest) {
     console.log(`❤️ API Route: GET /api/posts ${req.nextUrl.searchParams.toString()}`);
+
+    await sleep(5000);
 
     const postSearchParams = postSearch.load(req);
     const {items, total} = await searchHomePosts(postSearchParams);
